# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for advertisement store.",
        "title": "Swagger Advertisement store API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/SaenkoDmitry/advertisement-app"
        },
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:9100",
    "paths": {
        "/adv": {
            "get": {
                "description": "get all advertisements by params: orderBy, direction, page, limit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all advertisements by query params",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "number of rows",
                        "name": "limit",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "field order by: created_at, price",
                        "name": "orderBy",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "field direction: asc, desc",
                        "name": "direction",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AdvDto"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "create an advertisement by params: name, price, avatar, photos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create an advertisement",
                "parameters": [
                    {
                        "description": "advertisement",
                        "name": "adv",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AdvCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "id of created advertisement",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.Error"
                        }
                    }
                }
            }
        },
        "/adv/{id}": {
            "get": {
                "description": "get an advertisement by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get an advertisement",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Adv ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Adv"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gin.Error": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "error"
                },
                "meta": {
                    "type": "object"
                },
                "type": {
                    "type": "ErrorType"
                }
            }
        },
        "models.Adv": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "photos": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "models.AdvCreateReq": {
            "type": "object",
            "required": [
                "avatar",
                "name",
                "photos",
                "price"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "models.AdvDto": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "photos": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        }
    }
}